Added reactstrap, form and rendering state
